feat!: reworked base class
The base class has been reworked with an ABC. Additionally, there is now
a debug function – if you run a day's file with `--debug`, it looks for
a file named `example.txt` in the CWD, and loads that as your input.

Additionally, the base class now assumes that there is a single file per
day, instead of one per part. This was done in order to more easily
inherit Part A from Part B without resorting to the hacky tricks
previously used. Some small changes were necessary for interaction with
advent-of-code-data to download the correct file, but those should be
transparent to the user.

Finally, there was incorrect logic in determining whether to overwrite
files when templating – this has been fixed.